refactor(desktop): isolate renderer theme bootstrap
